address review comments:
- rm extraneous User model
- mv addresses to config dir
- clean up stake utils
