+++ packages/frontend/src/config/addresses.ts
@@ -65,6 +65,25 @@ if (process.env.NODE_ENV !== 'test') {
 const blocknativeDappid = process.env.REACT_APP_BNC_DAPP_ID
 const fortmaticApiKey = process.env.REACT_APP_FORTMATIC_KEY
 
+const stakingRewardsContracts = {
+  polygon: {
+    ETH: '0x7bCeDA1Db99D64F25eFA279BB11CE48E15Fda427',
+    MATIC: '0x7dEEbCaD1416110022F444B03aEb1D20eB4Ea53f',
+    USDC: '0x2C2Ab81Cf235e86374468b387e241DF22459A265',
+    USDT: '0x07932e9A5AB8800922B2688FB1FA0DAAd8341772',
+  },
+  gnosis: {
+    DAI: '0x12a3a66720dD925fa93f7C895bC20Ca9560AdFe7',
+    ETH: '0xC61bA16e864eFbd06a9fe30Aab39D18B8F63710a',
+    USDC: '0x5D13179c5fa40b87D53Ff67ca26245D3D5B2F872',
+  },
+}
+
+const rewardTokenAddresses = {
+  WMATIC: '0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270',
+  GNO: '0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb',
+}
+
 export {
   addresses,
   reactAppNetwork,
@@ -73,4 +92,6 @@ export {
   isMainnet,
   blocknativeDappid,
   fortmaticApiKey,
+  stakingRewardsContracts,
+  rewardTokenAddresses,
 }
