+++ packages/frontend/src/pages/Stake/Stake.tsx
@@ -7,7 +7,7 @@ import { useWeb3Context } from 'src/contexts/Web3Context'
 import StakeWidget from 'src/pages/Stake/StakeWidget'
 import useAsyncMemo from 'src/hooks/useAsyncMemo'
 import { useSelectedNetwork } from 'src/hooks'
-import { isMainnet } from 'src/config'
+import { isMainnet, rewardTokenAddresses, stakingRewardsContracts } from 'src/config'
 import { Div, Flex } from 'src/components/ui'
 import { findMatchingBridge, findNetworkBySlug, getNetworkProviderOrDefault } from 'src/utils'
 import { StakingRewards__factory } from '@hop-protocol/core/contracts'
@@ -15,31 +15,12 @@ import { providers, Signer } from 'ethers'
 import Network from 'src/models/Network'
 import { RaisedNetworkSelector } from 'src/components/NetworkSelector/RaisedNetworkSelector'
 
-const stakingRewardsContracts = {
-  polygon: {
-    ETH: '0x7bCeDA1Db99D64F25eFA279BB11CE48E15Fda427',
-    MATIC: '0x7dEEbCaD1416110022F444B03aEb1D20eB4Ea53f',
-    USDC: '0x2C2Ab81Cf235e86374468b387e241DF22459A265',
-    USDT: '0x07932e9A5AB8800922B2688FB1FA0DAAd8341772',
-  },
-  gnosis: {
-    DAI: '0x12a3a66720dD925fa93f7C895bC20Ca9560AdFe7',
-    ETH: '0xC61bA16e864eFbd06a9fe30Aab39D18B8F63710a',
-    USDC: '0x5D13179c5fa40b87D53Ff67ca26245D3D5B2F872',
-  },
-}
-
-const rewardTokenAddresses = {
-  WMATIC: '0x0d500b1d8e8ef31e21c99d1db9a6444d3adf1270',
-  GNO: '0x9C58BAcC331c9aa871AFD802DB6379a98e80CEdb',
-}
-
 const Stake: FC = () => {
-  const { bridges, sdk, networks, user, tokens } = useApp()
+  const { bridges, sdk, networks, tokens } = useApp()
   const { provider } = useWeb3Context()
 
   const availableNetworks = networks.filter(n =>
-    [ChainSlug.Gnosis, ChainSlug.Polygon].includes(n.slug as ChainSlug)
+    Object.keys(stakingRewardsContracts).includes(n.slug as ChainSlug)
   )
   const polygonNetwork = useMemo(() => findNetworkBySlug(ChainSlug.Polygon) as Network, [networks])
   const gnosisNetwork = useMemo(() => findNetworkBySlug(ChainSlug.Gnosis) as Network, [networks])
@@ -75,7 +56,7 @@ const Stake: FC = () => {
     const polygonProvider = await sdk.getSignerOrProvider(ChainSlug.Polygon)
     const _provider = getNetworkProviderOrDefault(ChainId.Polygon, polygonProvider, provider)
     return StakingRewards__factory.connect(stakingRewardsContracts.polygon.MATIC, _provider)
-  }, [sdk, provider, user])
+  }, [sdk, provider])
 
   // USDT
   const usdtBridge = useAsyncMemo(async () => findMatchingBridge(bridges, 'USDT'), [bridges])
@@ -86,7 +67,7 @@ const Stake: FC = () => {
     const polygonProvider = await sdk.getSignerOrProvider(ChainSlug.Polygon)
     const _provider = getNetworkProviderOrDefault(ChainId.Polygon, polygonProvider, provider)
     return StakingRewards__factory.connect(stakingRewardsContracts.polygon.USDT, _provider)
-  }, [sdk, provider, user])
+  }, [sdk, provider])
 
   // GNO
   const daiBridge = useAsyncMemo(async () => findMatchingBridge(bridges, 'DAI'), [bridges])
@@ -97,7 +78,7 @@ const Stake: FC = () => {
     const gnosisProvider = await sdk.getSignerOrProvider(ChainSlug.Gnosis)
     const _provider = getNetworkProviderOrDefault(ChainId.Gnosis, gnosisProvider, provider)
     return StakingRewards__factory.connect(stakingRewardsContracts.gnosis.DAI, _provider)
-  }, [sdk, provider, user])
+  }, [sdk, provider])
 
   // USDC
   const usdcBridge = useAsyncMemo(async () => findMatchingBridge(bridges, 'USDC'), [bridges])
